// src/app/graphql/types.generated.ts
export type Scalars = {
  ID: { input: string; output: ID; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.This scalar is serialized to a string in ISO 8601 format and parsed from a string in ISO 8601 format. */
  DateTimeISO: { input: Date; output: Date; }
  /** A field whose value conforms to the standard internet email address format as specified in HTML Spec: https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address. */
  EmailAddress: { input: Email; output: Email; }
  /** Represents NULL values */
  Void: { input: void; output: void; }
};
// ...
export type MutationTodoCreateArgs = {
  description: Scalars['String']['input'];
  title: Scalars['String']['input'];
};
// ...
export type EmailAddress = Scalars["EmailAddress"];
export type Void = Scalars["Void"];